add explicit checks for empty buffer and corresp. unittests; move old net5 tests into standard file
