+++ UnitTest/TestComponentCSharp_Tests.cs
@@ -8,6 +8,7 @@ using WinRT;
 
 using Windows.Foundation;
 using Windows.UI;
+using Windows.Storage;
 using Windows.Storage.Streams;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls.Primitives;
@@ -25,6 +26,7 @@ using System.Runtime.InteropServices.WindowsRuntime;
 using Windows.Security.Cryptography;
 using Windows.Security.Cryptography.Core;
 
+
 #if NET5_0
 using WeakRefNS = System;
 #else
@@ -40,6 +42,23 @@ namespace UnitTest
         public TestCSharp()
         {
             TestObject = new Class();
+        }
+
+        [Fact]
+        public void TestEmptyBufferToArray()
+        {
+            var buffer = new Windows.Storage.Streams.Buffer(0);
+            var array = buffer.ToArray();
+            Assert.True(array.Length == 0);
+        }
+
+        [Fact]
+        public void TestEmptyBufferCopyTo()
+        { 
+            var buffer = new Windows.Storage.Streams.Buffer(0);
+            byte[] array = { };
+            buffer.CopyTo(array);
+            Assert.True(array.Length == 0);
         }
 
         [Fact]
@@ -61,6 +80,37 @@ namespace UnitTest
         }
 
 #if NET5_0
+        [Fact]
+        public void TestAsStream()
+        {
+            using InMemoryRandomAccessStream winrtStream = new InMemoryRandomAccessStream();
+            using Stream normalStream = winrtStream.AsStream();
+            using var memoryStream = new MemoryStream();
+            normalStream.CopyTo(memoryStream);
+        }
+
+        async Task InvokeStreamWriteAndReadAsync()
+        {
+            var random = new Random(42);
+            byte[] data = new byte[256];
+            random.NextBytes(data);
+
+            using var stream = new InMemoryRandomAccessStream().AsStream();
+            await stream.WriteAsync(data, 0, data.Length);
+            stream.Seek(0, SeekOrigin.Begin);
+
+            byte[] read = new byte[256];
+            await stream.ReadAsync(read, 0, read.Length);
+            Assert.Equal(read, data);
+        }
+
+        [Fact]
+        public void TestStreamWriteAndRead()
+        {
+            Assert.True(InvokeStreamWriteAndReadAsync().Wait(1000));
+        }
+
+
         [Fact]
         public void TestDynamicInterfaceCastingOnValidInterface()
         {
